Filedropzone validators cleanup
